Revert "Added support for distributed exclusive saga locks via MySQL tables (part deux)"
